+++ R/cloneRate-package.R
@@ -1,0 +1,16 @@
+#' The 'cloneRate' package.
+#'
+#' @description Quickly estimate the net growth rate of a population or clone whose growth can be approximated by a birth-death branching process.
+#'
+#' @docType package
+#' @name cloneRate-package
+#' @aliases cloneRate
+#' @useDynLib cloneRate, .registration = TRUE
+#' @import methods
+#' @import Rcpp
+#' @import rstantools
+#' @importFrom rstan sampling
+#'
+#' @references Johnson et al. 2023 (under review at Bioinformatics). bioRxiv preprint <doi:10.1101/2023.02.24.529817>
+#'
+NULL