+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## Update for CRAN submission
 
+* Make birth-death MCMC available to all users of package
+* Move rstan from suggests to imports
+* Update rstan methods to be consistent with rstantools recommendations
 * Change installation instructions in README
 * Update title and description
 * Small changes to function documentation